Initial work on implementing Facebook session (issue #12)
Imported Facebook-SDK library and made some initial progress towards implementing
a Facebook session. Plugging in an active token into the
facebook_qube/session.defaults fbData and creating a new Facebook session
successfully logs in the user and speaks their name.

From here in we'll have to do the whole web log in thing as we have to ask for
a hefty set of extra permissions as Facebook is heavily locked down